+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     are overwritten by the ones provided on CLI.
     - Change banner logging to stderr instead of stdout.
     - New CLI flag `-or` to avoid creating result files if we didn't get any. 
+    - New CLI flag `-input-shell` to set the shell to be used by `input-cmd`
 
   - Changed
     - Pre-flight errors are now displayed also after the usage text to prevent the need to scroll through backlog.